+++ composeApp/src/commonMain/kotlin/proj/memorchess/axl/ui/pages/Explore.kt
@@ -1,12 +1,42 @@
 package proj.memorchess.axl.ui.pages
 
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.BoxWithConstraints
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
 import androidx.compose.runtime.*
+import androidx.compose.runtime.mutableStateListOf
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.unit.dp
-import proj.memorchess.axl.ui.components.board.control.ControllableBoard
+import compose.icons.FeatherIcons
+import compose.icons.feathericons.Save
+import compose.icons.feathericons.Trash
+import kotlinx.coroutines.launch
+import memorchess.composeapp.generated.resources.Res
+import memorchess.composeapp.generated.resources.description_board_next_move
+import org.jetbrains.compose.resources.stringResource
+import proj.memorchess.axl.core.engine.Game
+import proj.memorchess.axl.core.engine.pieces.vectors.King
+import proj.memorchess.axl.core.graph.nodes.NodeManager
+import proj.memorchess.axl.core.interactions.LinesExplorer
+import proj.memorchess.axl.ui.components.board.Board
+import proj.memorchess.axl.ui.components.board.Piece
+import proj.memorchess.axl.ui.components.board.StateIndicator
+import proj.memorchess.axl.ui.components.board.control.ControlButton
+import proj.memorchess.axl.ui.components.loading.LoadingWidget
+import proj.memorchess.axl.ui.layout.explore.ExploreLayoutContent
+import proj.memorchess.axl.ui.layout.explore.LandscapeExploreLayout
+import proj.memorchess.axl.ui.layout.explore.PortraitExploreLayout
 import proj.memorchess.axl.ui.pages.navigation.Destination
 
 @Composable
@@ -18,6 +48,97 @@ fun Explore() {
         .testTag(Destination.EXPLORE.name),
     horizontalAlignment = Alignment.CenterHorizontally,
   ) {
-    ControllableBoard(modifier = Modifier.fillMaxWidth())
+    LoadingWidget({ NodeManager.resetCacheFromDataBase() }) {
+      val modifier = Modifier.fillMaxWidth()
+      var inverted by remember { mutableStateOf(false) }
+      val linesExplorer = remember { LinesExplorer() }
+      val coroutineScope = rememberCoroutineScope()
+      val nextMoves = remember {
+        mutableStateListOf(*linesExplorer.getNextMoves().toTypedArray<String>())
+      }
+      remember {
+        linesExplorer.registerCallBack {
+          nextMoves.clear()
+          nextMoves.addAll(linesExplorer.getNextMoves())
+        }
+      }
+      val content = remember {
+        ExploreLayoutContent(
+          resetButton = { ControlButton.RESET.render(it) { linesExplorer.reset() } },
+          reverseButton = { ControlButton.REVERSE.render(it) { inverted = !inverted } },
+          backButton = { ControlButton.BACK.render(it) { linesExplorer.back() } },
+          forwardButton = { ControlButton.FORWARD.render(it) { linesExplorer.forward() } },
+          nextMoveButtons = {
+            if (nextMoves.isEmpty()) listOf<@Composable (() -> Unit)>({ NoNextMove() })
+            else
+              nextMoves.map<String, @Composable (() -> Unit)> {
+                { NextMoveButton(it) { coroutineScope.launch { linesExplorer.playMove(it) } } }
+              }
+          },
+          playerTurnIndicator = {
+            var playerTurn by remember {
+              mutableStateOf(linesExplorer.game.position.playerTurn == Game.Player.WHITE)
+            }
+            linesExplorer.registerCallBack {
+              playerTurn = linesExplorer.game.position.playerTurn == Game.Player.WHITE
+            }
+            Piece(if (playerTurn) King.white() else King.black())
+          },
+          stateIndicators = { StateIndicator(it, linesExplorer.state) },
+          saveButton = {
+            Button(
+              onClick = { coroutineScope.launch { linesExplorer.save() } },
+              it,
+              colors =
+                ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.primary),
+            ) {
+              Icon(FeatherIcons.Save, contentDescription = "Save")
+            }
+          },
+          deleteButton = {
+            Button(
+              onClick = { coroutineScope.launch { linesExplorer.delete() } },
+              it,
+              colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.error),
+            ) {
+              Icon(FeatherIcons.Trash, contentDescription = "Delete")
+            }
+          },
+          board = { Board(inverted, linesExplorer, it) },
+        )
+      }
+      BoxWithConstraints {
+        if (maxHeight > maxWidth) PortraitExploreLayout(modifier, content)
+        else LandscapeExploreLayout(modifier, content)
+      }
+    }
+  }
+}
+
+@Composable
+private fun NextMoveButton(move: String, playMove: () -> Unit) {
+  Box(
+    modifier =
+      Modifier.fillMaxSize()
+        .background(MaterialTheme.colorScheme.primaryContainer)
+        .testTag(stringResource(Res.string.description_board_next_move, move))
+        .clickable { playMove() },
+    contentAlignment = Alignment.Center,
+  ) {
+    Text(move)
+  }
+}
+
+@Composable
+private fun NoNextMove() {
+  Box(
+    modifier =
+      Modifier.fillMaxSize()
+        .background(MaterialTheme.colorScheme.primaryContainer)
+        .testTag("No next move")
+        .clickable {},
+    contentAlignment = Alignment.Center,
+  ) {
+    Text("No next move")
   }
 }
